requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
pandas>=2.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.0
fake-useragent>=1.4.0
retrying>=1.3.0
tqdm>=4.66.0
colorama>=0.4.0
rich>=13.7.0
click>=8.1.0
